# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "To Do List",
    "description": "To Do LIst REST API",
    "contact": {
      "email": "rmukunda.27@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "http://todobe-env.eba-bcfh3zx8.us-east-2.elasticbeanstalk.com",
  "basePath": "/",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/api/list/add": {
      "post": {
        "tags": ["Lists"],
        "summary": "Add a list for a user",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "listName",
            "description": "Name of the list created by the user",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Personal"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "List Added!!"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Request Body",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "List Name is missing in the request body"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/list/all": {
      "get": {
        "tags": ["Lists"],
        "summary": "Get all lists for a user",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "listName": {
                    "type": "string",
                    "example": "Personal"
                  },
                  "_id": {
                    "type": "string",
                    "example": "5fb694c0d884419aa8c5d33a"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/task/create": {
      "post": {
        "tags": ["Tasks"],
        "summary": "Create new task",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "title",
            "description": "Task Title",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Submit Project"
            }
          },
          {
            "in": "body",
            "name": "finishByDate",
            "description": "Task Finish Date",
            "required": true,
            "schema": {
              "type": "string",
              "example": "12/15/2020"
            }
          },
          {
            "in": "body",
            "name": "finishByTime",
            "description": "Task Finish Timer",
            "required": true,
            "schema": {
              "type": "string",
              "example": "06:00 PM"
            }
          },
          {
            "in": "body",
            "name": "important",
            "description": "Indicates if the task is important",
            "required": true,
            "schema": { "type": "boolean", "example": "false" }
          },
          {
            "in": "body",
            "name": "list",
            "description": "List to which the task belongs",
            "required": false,
            "schema": { "type": "string", "example": "Fitness" }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Task Created!!"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Request Body",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Some parameters are missing in the request body"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/task/all": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get all tasks for a user",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "pageSize",
            "description": "Number of tasks per page",
            "required": false,
            "schema": {
              "type": "number",
              "example": 15
            }
          },
          {
            "in": "query",
            "name": "currentPage",
            "description": "Current page index",
            "required": false,
            "schema": {
              "type": "number",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "object",
              "properties": {
                "tasks": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                },
                "totalTasks": {
                  "type": "number",
                  "example": 5
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/task/day": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get all tasks for a user for a given date",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "description": "Date for which the tasks will have to be retrieved",
            "required": true,
            "schema": {
              "type": "string",
              "example": "12/15/2020"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "description": "Number of tasks per page",
            "required": false,
            "schema": {
              "type": "number",
              "example": 15
            }
          },
          {
            "in": "query",
            "name": "currentPage",
            "description": "Current page index",
            "required": false,
            "schema": {
              "type": "number",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "object",
              "properties": {
                "tasks": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                },
                "totalTasks": {
                  "type": "number",
                  "example": 5
                }
              }
            }
          },
          "400": {
            "description": "Invalid Request Body",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Some query parameters are missing in the request"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/task/important": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get all important tasks for a user",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "pageSize",
            "description": "Number of tasks per page",
            "required": false,
            "schema": {
              "type": "number",
              "example": 15
            }
          },
          {
            "in": "query",
            "name": "currentPage",
            "description": "Current page index",
            "required": false,
            "schema": {
              "type": "number",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "object",
              "properties": {
                "tasks": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                },
                "totalTasks": {
                  "type": "number",
                  "example": 5
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/task/list": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get all tasks for a user",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "list",
            "description": "Date for which the tasks will have to be retrieved",
            "required": true,
            "schema": {
              "type": "string",
              "example": "fitness"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "description": "Number of tasks per page",
            "required": false,
            "schema": {
              "type": "number",
              "example": 15
            }
          },
          {
            "in": "query",
            "name": "currentPage",
            "description": "Current page index",
            "required": false,
            "schema": {
              "type": "number",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "object",
              "properties": {
                "tasks": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                },
                "totalTasks": {
                  "type": "number",
                  "example": 5
                }
              }
            }
          },
          "400": {
            "description": "Invalid Request Body",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Some query parameters are missing in the request"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/task/:id": {
      "put": {
        "tags": ["Tasks"],
        "summary": "Update the completion status of a task",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "completed",
            "description": "Task completion status",
            "required": true,
            "schema": {
              "type": "boolean",
              "example": "true"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Task status change successful!"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Request Body",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Some parameters are missing in the request"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/task/:taskid": {
      "delete": {
        "tags": ["Tasks"],
        "summary": "Delete task of a user",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Deletion successful!"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    }
  },
...
}